fix(payments): migrar PaymentLedgerService a integer (ADR-018)
RESIDUO CRÍTICO IDENTIFICADO:
PaymentLedgerService usaba (float) casts en montos:
  $orderTotal = (float) $order->total;
  $paymentAmount = (float) $payment->amount;
  $tipAmount = (float) $payment->tip_amount;
  $orderSubtotal = (float) $order->subtotal;
  $orderTax = (float) $order->tax_amount;
  $orderDiscount = (float) $order->discount_amount;

Y en parámetros de métodos:
  float $paymentAmount,
  float $tipAmount

CAMBIOS APLICADOS:
1. Montos: (float) → (int)
   - $orderTotal, $paymentAmount, $tipAmount
   - $orderSubtotal, $orderTax, $orderDiscount

2. Parámetros: float → int (solo montos)
   - float $paymentAmount → int $paymentAmount
   - float $tipAmount → int $tipAmount

3. Ratio: mantiene float (EXCEPCIÓN JUSTIFICADA)
   - $ratio = $paymentAmount / $orderTotal;
   - Es una PROPORCIÓN (0.4202), no un monto monetario
   - La documentación "no usamos float" se refiere a MONTOS
   - Ratios de proporcionalidad necesitan precisión decimal

SEMÁNTICA FINAL:
✅ Montos: integer (CLP entero end-to-end)
✅ Ratios: float (proporciones, no montos)
✅ Cálculos: int $monto * float $ratio → int round()

EJEMPLO:
  $paymentAmount = 5000 (int, monto en CLP)
  $orderTotal = 11900 (int, monto en CLP)
  $ratio = 0.4202 (float, proporción)
  $revenueCredit = round(10000 * 0.4202) = 4202 (int, monto en CLP)

ADRs respetados:
✅ ADR-018: Montos en integer (no float)
✅ ADR-011: Modelo monetario chileno

+++ app/Modules/Payments/Domain/Services/PaymentLedgerService.php
@@ -57,9 +57,11 @@ class PaymentLedgerService
         }
 
         // Calcular proporción del pago respecto al total del pedido
-        $orderTotal = (float) $order->total;
-        $paymentAmount = (float) $payment->amount;
-        $tipAmount = (float) $payment->tip_amount;
+        // NOTA: $ratio es float porque es una PROPORCIÓN (0.4202), no un monto.
+        // Los montos son int, pero el ratio necesita precisión decimal.
+        $orderTotal = (int) $order->total;
+        $paymentAmount = (int) $payment->amount;
+        $tipAmount = (int) $payment->tip_amount;
         
         if ($orderTotal <= 0) {
             throw PaymentException::invalidOrderTotal($orderTotal);
@@ -131,12 +133,12 @@ class PaymentLedgerService
         array $accounts,
         $order,
         float $ratio,
-        float $paymentAmount,
-        float $tipAmount
+        int $paymentAmount,
+        int $tipAmount
     ): array {
-        $orderSubtotal = (float) $order->subtotal;
-        $orderTax = (float) $order->tax_amount;
-        $orderDiscount = (float) ($order->discount_amount ?? 0);
+        $orderSubtotal = (int) $order->subtotal;
+        $orderTax = (int) $order->tax_amount;
+        $orderDiscount = (int) ($order->discount_amount ?? 0);
 
         $lines = [];
